+++ .github/actions/run-tests/tests/entrypoints/default-entrypoint.sh
@@ -1,16 +1,14 @@
 #! /bin/bash
 
+/entrypoints/minimal-default-entrypoint.sh
+
 if [ `whoami` = root ]; then
 	
-	if [ -z "$QUICK_MODE" -o "$QUCK_MODE" = n ]; then
-		echo "Setting uid and gid to $RUNNER_UID/$RUNNER_GID"
-		usermod -u $RUNNER_UID runner
-		groupmod -g $RUNNER_GID runner
-		
+	if [ "$QUCK_MODE" = y ]; then
+		echo "Quick mode activated. No permission update is carried out"
+	else
 		echo "Changing ownership of files to runner"
 		chown -R runner: /nextcloud
-	else
-		echo "Quick mode activated. No permission update is carried out"
 	fi
 	
 	if [ -n "$DEBUG_MODE" ]; then
@@ -36,6 +34,10 @@ if [ `whoami` = root ]; then
 			DEBUG_TRACE_FORMAT=1
 		fi
 		
+		if [ -z "$XDEBUG_LOG_LEVEL" ]; then
+			XDEBUG_LOG_LEVEL=3
+		fi
+		
 		mode="develop,coverage,$DEBUG_MODE"
 		
 		cat >> /usr/local/etc/php/conf.d/docker-php-ext-xdebug.ini <<- EOF
@@ -46,6 +48,7 @@ if [ `whoami` = root ]; then
 			xdebug.client_port = $DEBUG_PORT
 			xdebug.client_host = $DEBUG_HOST
 			xdebug.trace_format = $DEBUG_TRACE_FORMAT
+			xdebug.log_level = $XDEBUG_LOG_LEVEL
 			EOF
 		
 	fi