+++ .github/actions/run-tests/tests/install-sudoers.sh
@@ -1,0 +1,3 @@
+#!/bin/bash -e
+
+echo 'runner ALL=(ALL) NOPASSWD: ALL' >> /etc/sudoers